Added Htmlable Support For TableHeader.
Always making view, and pass it as table header is not something we want all time as we can use htmlstring in filament.

So I add support for this.

Also i see there may be issue with table headeraction, So what i have done , If the table header pass as argument is Illuminate\Support\HtmlString, I just added the action with the heading.