Update 05-validation.md
The substitution field `:attribute` used in a custom validation rule message will honour the `validationAttribute()` if set. In the original example `{$attribute}` will substitute the name of the field exactly as supplied in the `->make('name')` method, and so will usually be wrong (e.g. "first_name" instead of "First Name").