Fixed: Addressed $loop->iteration placement outside loops and corrected minor structural issues.
When using limitedRemainingText and no remaining image is present, we currently display +0, which is not ideal. Instead, if the number of images exceeds the set limit, for example, if we have 5 images and the limit is set to 3, we should display +2. Additionally, to address a performance issue, we are currently using count($limitedState) in 6 instances and count($state) in 2 instances. To avoid duplication, I have created separate variables.

Note: We have $loop->iteration in 2 lines of code, but these lines are not within a loop. Additionally, there is a minor structural problem which has been corrected as expected.